requirements.txt
fastapi
uvicorn
redis
python-dotenv
pyotp
smartapi-python
asyncio
logzero
websocket-client
